+++ main/ActiveGithub.py
@@ -8,8 +8,10 @@ THISDIR = str(Path(__file__).resolve().parent)
 sys.path.insert(0, os.path.dirname(THISDIR) + "/lib")
 
 from metprint import LogType
+#pylint: disable=import-error
 import githubGraph
 from utils import printf, getUsernameAndLifespan
+#pylint: enable=import-error
 
 _username, death = getUsernameAndLifespan()
 author, repoName = input(